feat(interfaces): change room ID type from int to long in repository and DTOs

+++ Backend/app/Core/Interfaces/IRoomReadModelRepository.cs
@@ -5,5 +5,5 @@ namespace Backend.app.Core.Interfaces;
 public interface IRoomReadModelRepository
 {
     Task<IEnumerable<RoomDetailModel>> GetAllRoomDetailsAsync();
-    Task<RoomDetailModel?> GetRoomDetailByIdAsync(int roomId);
+    Task<RoomDetailModel?> GetRoomDetailByIdAsync(long roomId);
 }
